requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
plotly==5.18.0
Flask==3.0.2
Werkzeug==3.0.1
pandas==2.2.1
numpy==1.26.3
dash-ag-grid==31.0.1
pyarrow==14.0.2
shapely==2.0.2
python-dotenv>=1.0.1
bcrypt>=4.0.1